+++ asfsmd.py
@@ -35,7 +35,7 @@ import requests
 import asf_search as asf
 
 
-__version__ = "1.0.0"
+__version__ = "1.1.0"
 __all__ = ["download_annotations", "main"]